arm: Fix FDT loaded in wrong location
For reasons not fully understood, the FDT messes up the boot if it is loaded
in the location that the default allocator returns. It is suspected that the
location is too close to the kernel location, and the kernel may be using
the space to decompress.

In any case, this patch fixes the problem.

See https://tracker.mender.io/browse/MEN-2404